+++ src/main/java/cart/controller/CartItemApiController.java
@@ -1,10 +1,10 @@
 package cart.controller;
 
-import cart.domain.Member;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemRequest;
-import cart.dto.CartItemResponse;
+import cart.controller.dto.CartItemResponse;
+import cart.domain.member.Member;
 import cart.service.CartItemService;
+import cart.service.dto.CartItemQuantityUpdateRequest;
+import cart.service.dto.CartItemRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
