+++ src/main/java/cart/controller/dto/CouponResponse.java
@@ -1,17 +1,19 @@
-package cart.dto;
+package cart.controller.dto;
 
-public class CouponTypeResponse {
+public class CouponResponse {
 
     private final Long id;
     private final String name;
     private final int discountAmount;
     private final String description;
+    private final boolean used;
 
-    public CouponTypeResponse(final Long id, final String name, final int discountAmount, final String description) {
+    public CouponResponse(final Long id, final String name, final int discountAmount, final String description, final boolean used) {
         this.id = id;
         this.name = name;
         this.discountAmount = discountAmount;
         this.description = description;
+        this.used = used;
     }
 
     public Long getId() {
@@ -28,5 +30,9 @@ public class CouponTypeResponse {
 
     public String getDescription() {
         return description;
+    }
+
+    public boolean isUsed() {
+        return used;
     }
 }
