+++ src/main/java/cart/dao/CartItemDao.java
@@ -1,9 +1,9 @@
 package cart.dao;
 
-import cart.dao.entity.CartItemResultMap;
-import cart.domain.CartItem;
-import cart.domain.Member;
+import cart.dao.dto.CartItemResultMap;
 import cart.domain.Product;
+import cart.domain.cart.CartItem;
+import cart.domain.member.Member;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
