+++ src/main/java/cart/domain/member/MemberRepository.java
@@ -1,6 +1,4 @@
-package cart.repository;
-
-import cart.domain.Member;
+package cart.domain.member;
 
 public interface MemberRepository {
