+++ src/main/java/cart/domain/order/OrderRepository.java
@@ -1,6 +1,4 @@
-package cart.repository;
-
-import cart.domain.Order;
+package cart.domain.order;
 
 import java.util.List;
