+++ src/main/java/cart/service/dto/CartItemRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.service.dto;
 
 public class CartItemRequest {
     private Long productId;
