+++ src/main/java/cart/service/dto/CouponReissueRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.service.dto;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
