+++ src/main/java/cart/service/dto/ProductRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.service.dto;
 
 public class ProductRequest {
     private String name;
