+++ src/test/java/cart/acceptance/AcceptanceTest.java
@@ -1,7 +1,7 @@
 package cart.acceptance;
 
 import cart.dao.MemberDao;
-import cart.domain.Member;
+import cart.domain.member.Member;
 import io.restassured.RestAssured;
 import io.restassured.specification.RequestSpecification;
 import org.junit.jupiter.api.BeforeEach;
