+++ src/test/java/cart/repository/OrderJdbcRepositoryTest.java
@@ -3,11 +3,12 @@ package cart.repository;
 import cart.dao.OrderDao;
 import cart.dao.OrderItemDao;
 import cart.dao.ProductDao;
-import cart.domain.Member;
-import cart.domain.Order;
-import cart.domain.OrderItem;
 import cart.domain.Product;
 import cart.domain.coupon.Coupon;
+import cart.domain.member.Member;
+import cart.domain.order.Order;
+import cart.domain.order.OrderItem;
+import cart.domain.order.OrderRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
